feat: refactor cookie handling and improve UI components for better user experience
