+++ src/app/[locale]/(auth)/page.jsx
@@ -39,12 +39,12 @@ import VideoBackground from "@/components/ui/VideoBackground";
 import VideoPlayer from "@/components/ui/VideoPlayer";
 import { API_BASE_URL } from "@/config/settings";
 import { useUser } from "@/hooks/data/useUser";
-import { useCookies } from "@/store/auth/signup";
+import { useCookiesStore } from "@/store/auth/cookiesStore";
 
 function Home() {
   const [showDonatePopup, setShowDonatePopup] = useState(false);
   const { user } = useUser();
-  const { cookieIsAccepted } = useCookies();
+  const { cookieIsAccepted } = useCookiesStore;
 
   const [showPdf, setShowPdf] = useState(false);
