+++ src/app/[locale]/error.jsx
@@ -1,15 +1,13 @@
 "use client";
 
 import Link from "next/link";
-import { useEffect } from "react";
-import { IoHomeOutline } from "react-icons/io5";
-import { IoRefreshOutline } from "react-icons/io5";
+import { LuHouse, LuRotateCcw } from "react-icons/lu";
+
+import { Contact } from "@/components/layout/Contact";
 
 export default function Error({ error }) {
-
   return (
-    <main className="flex flex-col items-center justify-center h-screen w-screen overflow-hidden">
-
+    <main className="flex flex-col items-center justify-center h-full w-screen overflow-hidden">
       <div className="space-y-8 text-center">
         <div className="relative">
           <div className="w-24 h-24 bg-red-100 rounded-full mx-auto flex items-center justify-center">
@@ -30,41 +28,23 @@ export default function Error({ error }) {
         <div className="flex flex-col sm:flex-row gap-4 justify-center items-center">
           <button
             onClick={() => window.location.reload()}
-            className="flex items-center gap-2 px-6 py-3 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors"
+            className="flex items-center justify-center w-44 gap-2 px-2 py-2 cursor-pointer hover:scale-105 transition-transform ease-in-out duration-300 bg-[#001840] text-white rounded-lg hover:bg-[#00840]/50 overflow-hidden"
           >
-            <IoRefreshOutline className="w-4 h-4" />
-            Try Again
+            <LuRotateCcw className="w-4 h-4" />
+            <> Try Again</>
           </button>
 
           <Link
             href="/"
-            className="flex items-center gap-2 px-6 py-3 bg-white text-gray-600 rounded-lg border border-gray-200 hover:bg-gray-50 transition-colors"
+            className="flex items-center justify-center w-44 gap-2 px-2 py-2 cursor-pointer hover:scale-105 transition-transform ease-in-out duration-300 text-[#001840] rounded-lg border border-[#001840] overflow-hidden"
           >
-            <IoHomeOutline className="w-4 h-4" />
+            <LuHouse className="w-4 h-4" />
             Back to Home
           </Link>
         </div>
 
-        {/* {process.env.NODE_ENV === "development" && (
-          <div className="mt-8">
-            <details className="text-left bg-white p-4 rounded-lg border border-gray-200">
-              <summary className="cursor-pointer text-sm text-gray-500 hover:text-gray-700">
-                Error Details
-              </summary>
-              <pre className="mt-2 text-xs text-gray-600 overflow-auto p-2 bg-gray-50 rounded">
-                {error.stack || error.message}
-              </pre>
-            </details>
-          </div>
-        )} */}
-
-        <div className="text-sm text-gray-500">
-          <p>
-            Need help?{" "}
-            <a href="/contact" className="text-blue-600 hover:underline ml-1">
-              Contact Support
-            </a>
-          </p>
+        <div className="text-sm">
+          <Contact />
           {error.digest && (
             <p className="mt-2">
               Error ID:{" "}
